+++ Sources/TuistAnalytics/TuistAnalytics.swift
@@ -1,9 +1,19 @@
 import Foundation
+import TSCBasic
 import TuistAsyncQueue
+import TuistGraph
+import TuistLoader
 
 public final class TuistAnalytics {
-    public static func bootstrap() {
-        AsyncQueue.sharedInstance.register(dispatcher: TuistAnalyticsDispatcher())
+    public static func bootstrap(configLoader: ConfigLoader = ConfigLoader(manifestLoader: ManifestLoader())) throws {
+        let path: AbsolutePath
+        if let argumentIndex = CommandLine.arguments.firstIndex(of: "--path") {
+            path = AbsolutePath(CommandLine.arguments[argumentIndex], relativeTo: .current)
+        } else {
+            path = .current
+        }
+
+        AsyncQueue.sharedInstance.register(dispatcher: TuistAnalyticsDispatcher(cloud: try configLoader.loadConfig(path: path).cloud))
         AsyncQueue.sharedInstance.start() // Re-try to send all events that got persisted and haven't been sent yet
     }
 }