+++ Sources/TuistGraph/Models/Cloud.swift
@@ -4,6 +4,7 @@ import Foundation
 public struct Cloud: Equatable, Hashable {
     /// Cloud option.
     public enum Option: String, Codable, Equatable {
+        case analytics
         case insights
     }