+++ Sources/tuist/main.swift
@@ -6,7 +6,7 @@ import TuistSupport
 if CommandLine.arguments.contains("--verbose") { try? ProcessEnv.setVar(Constants.EnvironmentVariables.verbose, value: "true") }
 
 TuistSupport.LogOutput.bootstrap()
-TuistAnalytics.bootstrap()
+try TuistAnalytics.bootstrap()
 
 import TuistKit